Salesanalyst

+++ lib/item_repository.rb
@@ -28,7 +28,7 @@ class ItemRepository
 
   def find_all_by_price_in_range(range)
     @all.find_all do |obj|
-      range.include?(obj.unit_price.to_i)
+      range.include?(obj.unit_price.to_f)
     end
   end
